// src/app/share/wy-ui/wy-slider/wy-slider-types.ts
import { Observable } from 'rxjs'

// 定义style属性
export type WySliderStyle = {
    width?: string | null;
    height?: string | null;
    left?: string | null;
    bottom?: string | null;
}

export type SliderEventObserverConfig = {
    start: string;
    move: string;
    end: string;
    filter: (e: Event) => boolean;
    pluckKey: string[];
    startPlucked$?:Observable<number>;
    moveResolved$?:Observable<number>;
    end$?:Observable<Event>;
}

export type SliderValue = number | null;